webnfc/manual-test/index.html

<meta charset=utf-8>
<title>Web NFC Watch Tests</title>
<link rel="author" title="Intel" href="http://www.intel.com"/>
<link rel="author" title="Zhao Ming" href="mailto:mingx.zhao@intel.com"/>
<link rel="help" href="https://w3c.github.io/web-nfc/#the-nfc-interface"/>
<script src="./resources/testharness.js"></script>
<script src="./resources/testharnessreport.js"></script>
<script src="./nfc_help.js"></script>

<div id="log"></div>
<script>
//  navigator.nfc.watch((message) => {
//    if (message.data[0].recordType == 'empty') {
//      console.log("DEBUG-1")
//      navigator.nfc.push({
//        url: "/custom/path",
//        data: [{ recordType: "text", data: 'Hello World' }]
//      });
//    } else {
//      console.log("DEBUG-2")
//      console.log('Read message written by ' + message.url);
//      processMessage(message);
//    }
//  }).then(() => {
//    console.log("DEBUG-3")
//    console.log("Added a watch.");
//  }).catch((error) => {
//    console.log("DEBUG-4")
//    console.log("Adding watch failed: " + error.name);
//  });
//  
//  function processMessage(message) {
//    console.log("DEBUG-5")
//    console.log('-----------------------------')
//    for (let record of message.data) {
//      console.log("DEBUG-6")
//      switch (record.recordType) {
//        case "text":
//          console.log("DEBUG-7")
//          console.log('Data is text: ' + record.data);
//          break;
//        case "url":
//          console.log('Data is URL: ' + record.data);
//          break;
//        case "json":
//          console.log('JSON data: ' + record.data.myProperty.toString());
//          break;
//        case "opaque":
//          if (record.mediaType == 'image/png') {
//            var img = document.createElement("img");
//            img.src = URL.createObjectURL(new Blob(record.data, record.mediaType));
//            img.onload = function() {
//              window.URL.revokeObjectURL(this.src);
//            };
//          }
//          break;
//      }
//    }
//  }

//function pushCheck(arg1, arg2, arg3) {
  // arg1 : context    arg2 : type     arg3 : comment 
//  promise_test(t => {
//    return navigator.nfc.push("aaabbbccc")
//      .then(() => {
//        navigator.nfc.watch((message)
//            .then(()=> {
//              console.log(message.data.recordType)
//              assert_equals(message.data.recordType, "text");
//            })
//        )})
//  }, "test comm")
//}
//let a = "test strings zhaoming"
//let b = "1text"
//let c = "this is test case for automachie"


//promise_test(t => {
//  let watchoptions = {
//    recordType: "text",
//    mediaType: "text/plain"
//  }
// return navigator.nfc.push("1233333")
//   .then(() => {
//      return new Promise(resolve => {
//        navigator.nfc.watch((message) => resolve(message), watchoptions);
//        }).then((message) => {
//          console.log("in");
//           console.log("message", message)
//           for (let record of message.data) {
//            if (record.recordType == "text") {
//              assert_equals(record.recordType, "text");
//            }
//          }
//      })
//   })  
//}, "wanming test")
function checkPushType(pushMessage, types, watchOptions, comments) {
  //arg1: content, arg2: type, arg3:option, arg4: comment
  promise_test(t => {
    return navigator.nfc.push(pushMessage)
      .then(() => {
        return new Promise(resolve => {
          navigator.nfc.watch((message) => resolve(message), watchOptions);
        }).then((message) => {
          for (let record of message.data) {
            if (record.recordType) {
              assert_equals(record.recordType, types);
            }
          }
        })
      })
  }, comments)
}
checkPushType("9999", "text1", {recordType: "text"}, "this is ****")
</script>
